+++ src/api_calls/user_calls.rs
@@ -1,7 +1,7 @@
 use crate::{
     api_calls::check_id_slug,
-    request::request_rel,
-    structures::{user_structs::*, ID},
+    request::{request, API_URL_BASE},
+    structures::{project_structs::Project, user_structs::*},
     Ferinth, Result,
 };
 
@@ -10,73 +10,73 @@ impl Ferinth {
     ///
     /// Example:
     /// ```rust
+    /// # use ferinth::structures::user_structs::UserRole;
+    /// # #[tokio::main]
+    /// # async fn main() -> Result<(), ferinth::Error> {
     /// # let modrinth = ferinth::Ferinth::new("ferinth-example");
-    /// # tokio_test::block_on( async {
     /// let jellysquid = modrinth.get_user("TEZXhE2U").await?;
-    /// assert_eq!(
-    ///     jellysquid.role,
-    ///     ferinth::structures::user_structs::UserRole::Developer,
-    /// );
-    /// # Ok::<(), ferinth::Error>(())
-    /// # } );
+    /// assert!(jellysquid.role == UserRole::Developer);
+    /// # Ok(()) }
     /// ```
     pub async fn get_user(&self, user_id: &str) -> Result<User> {
         check_id_slug(user_id)?;
-        Ok(request_rel(self, format!("/user/{}", user_id))
+        Ok(request(self, API_URL_BASE.join("user/")?.join(user_id)?)
             .await?
             .json()
             .await?)
     }
 
-    /// Get a list of mod IDs of mods that the user owns
+    /// Get a list of projects that the user owns
     ///
     /// Example:
     /// ```rust
+    /// # #[tokio::main]
+    /// # async fn main() -> Result<(), ferinth::Error> {
     /// # let modrinth = ferinth::Ferinth::new("ferinth-example");
-    /// # tokio_test::block_on( async {
-    /// let jellysquid_mods = modrinth.list_mods("TEZXhE2U").await?;
-    /// assert!(vec![
-    ///         "hEOCdOgW".to_string(),
-    ///         "AANobbMI".to_string(),
-    ///         "gvQqBUqZ".to_string(),
-    ///         "AZomiSrC".to_string()
-    ///     ]
-    ///     .iter()
-    ///     // So that the order of the mods isn't checked
-    ///     .all(
-    ///         |mod_id| jellysquid_mods.contains(mod_id)
-    ///     )
-    /// );
-    /// # Ok::<(), ferinth::Error>(())
-    /// # } );
+    /// let jellysquid_projects = modrinth.list_projects("TEZXhE2U").await?;
+    /// assert!(jellysquid_projects.len() == 4);
+    /// # Ok(()) }
     /// ```
-    pub async fn list_mods(&self, user_id: &str) -> Result<Vec<ID>> {
+    pub async fn list_projects(&self, user_id: &str) -> Result<Vec<Project>> {
         check_id_slug(user_id)?;
-        Ok(request_rel(self, format!("/user/{}/mods", user_id))
-            .await?
-            .json()
-            .await?)
+        let mut user_id = user_id.to_string();
+        user_id.push('/');
+        Ok(request(
+            self,
+            API_URL_BASE
+                .join("user/")?
+                .join(&user_id)?
+                .join("projects")?,
+        )
+        .await?
+        .json()
+        .await?)
     }
 
     /// List the members of team with ID `team_id`
     ///
     /// Example:
     /// ```rust
+    /// # #[tokio::main]
+    /// # async fn main() -> Result<(), ferinth::Error> {
     /// # let modrinth = ferinth::Ferinth::new("ferinth-example");
-    /// # tokio_test::block_on( async {
     /// let mod_menu_team = modrinth.list_team_members("VMz4FpgB").await?;
-    /// assert_eq!(
-    ///     mod_menu_team.len(),
-    ///     4,
-    /// );
-    /// # Ok::<(), ferinth::Error>(())
-    /// # } );
+    /// assert!(mod_menu_team.len() == 4);
+    /// # Ok(()) }
     /// ```
     pub async fn list_team_members(&self, team_id: &str) -> Result<Vec<TeamMember>> {
         check_id_slug(team_id)?;
-        Ok(request_rel(self, format!("/team/{}/members", team_id))
-            .await?
-            .json()
-            .await?)
+        let mut team_id = team_id.to_string();
+        team_id.push('/');
+        Ok(request(
+            self,
+            API_URL_BASE
+                .join("team/")?
+                .join(&team_id)?
+                .join("members")?,
+        )
+        .await?
+        .json()
+        .await?)
     }
 }