+++ src/lib.rs
@@ -7,36 +7,33 @@
 //!
 //! This crate includes the following:
 //!
-//! - All structure definitions based on <https://github.com/modrinth/labrinth/wiki/API-Documentation#structure-definitions>
+//! - All structure definitions based on <https://docs.modrinth.com/api-spec/>
 //! - Includes the following API calls:
-//!   - Get mod by mod ID
-//!   - List mod categories
+//!   - Get project by project ID
+//!   - List project categories
 //!   - List mod loaders
 //!   - List game versions
 //!   - Get user by user ID
 //!   - List team members by team ID
 //!   - Get version by version ID
-//!   - Get version by file hash
-//!   - List versions by mod ID
+//!   - Get version by a version file's hash
+//!   - List versions by project ID
 //!   - Download version file
 //!
 //! URL traversal is blocked because all IDs are verified.
 //! ```
+//! # #[tokio::main]
+//! # async fn main() {
 //! # let modrinth = ferinth::Ferinth::new("ferinth-example");
-//! # tokio_test::block_on( async {
-//! assert!(modrinth.get_mod("sodium/version").await.is_err());
-//! # } );
+//! assert!(modrinth.get_project("sodium/version").await.is_err());
+//! # }
 //! ```
 //!
-//! This crate uses [Rustls](https://docs.rs/rustls/) rather than OpenSSL, because OpenSSL is outdated and slower.
+//! This crate uses [RusTLS](https://docs.rs/rustls/) rather than OpenSSL, because OpenSSL is outdated and slower.
 //!
-//! The following features have not yet been implemented
-//! - Search mods
+//! The following features still need to be implemented
+//! - Search projects
 //! - User authentication
-//! - Get current user (constrained by the lack of user authentication)
-//!
-//! Unfortunately, I am not planning to implement any of these features in this first version due to poor documentation.
-//! I will add these features in the next version, version 2, once the [Modrinth API v2](https://docs.modrinth.com/api-spec/) rolls out.
 //!
 //! ## Example (and Tutorial)
 //!
@@ -63,29 +60,30 @@
 //! let api = Ferinth::new("example");
 //!
 //! // Now, lets get the Sodium mod
-//! // You can use the mod ID, or the mod slug
-//! // The mod ID will never change but the mod slug can change at anytime
-//! // Using the mod slug
-//! let sodium = api.get_mod("sodium").await?;
-//! // Using the mod ID
-//! let sodium = api.get_mod("AANobbMI").await?;
+//! // You can use the project ID, or the project slug
+//! // The project ID will never change but the project slug can change at anytime
+//!
+//! // Using the project slug
+//! let sodium = api.get_project("sodium").await?;
+//! // Using the project ID
+//! let sodium = api.get_project("AANobbMI").await?;
 //!
 //! // Now lets get the versions that the Sodium mod has
 //! let sodium_versions = api.list_versions("AANobbMI").await?;
 //!
 //! // The versions are sorted chronologically so the first element should be the latest one
 //! let latest_version = &sodium_versions[0];
-//! // And now we can get this version's mod file, which is called a version file
+//! // And now we can get this version's mod JAR file, which is called a version file
 //! let version_file = &latest_version.files[0];
 //!
 //! // Then we can download this version file
 //! let contents = api.download_version_file(version_file).await?;
-//! // And next, lets open the file we want to write this to
+//! // And next, let's open the file we want to write this to
 //! let mut mod_file = File::create("Sodium.jar")?;
 //! // And finally, we can write the contents to mod_file
 //! mod_file.write_all(&contents)?;
 //!
-//! // Now you can use load the JAR file using a mod loader. To play Sodium. you should use Fabric
+//! // Now you can load the JAR file using a mod loader. To play Sodium, you need the Fabric Loader
 //! # Ok(())
 //! # }
 //! ```
@@ -102,6 +100,8 @@ pub enum Error {
     NotSHA1,
     #[error("{}", .0)]
     ReqwestError(#[from] reqwest::Error),
+    #[error("{}", .0)]
+    URLParseError(#[from] url::ParseError),
 }
 
 pub(crate) type Result<T> = std::result::Result<T, Error>;
@@ -113,7 +113,7 @@ pub(crate) type Result<T> = std::result::Result<T, Error>;
 /// # use ferinth::Ferinth;
 /// let modrinth = Ferinth::new("ferinth-example");
 /// // Use the instance to call the API
-/// let sodium_mod = modrinth.get_mod("sodium");
+/// let sodium_mod = modrinth.get_project("sodium");
 /// ```
 #[derive(Debug, Clone)]
 pub struct Ferinth {